rename activation module
